+++ src/app/heroes/pages/home/home.component.ts
@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
+import { Auth } from 'src/app/auth/pages/interfaces/auth.interface';
+import { AuthService } from 'src/app/auth/services/auth.service';
 
 @Component({
   selector: 'app-home',
@@ -11,9 +14,20 @@ import { Component, OnInit } from '@angular/core';
 })
 export class HomeComponent implements OnInit {
 
-  constructor() { }
+  get nombreUsuario() {
+    return this.authService.nombreUsuario;
+  }
+
+  constructor(
+    private router: Router,
+    private authService: AuthService
+  ) { }
 
   ngOnInit(): void {
+  }
+
+  logout() {
+    this.router.navigate(['./auth']);
   }
 
 }