+++ src/app/heroes/services/heroes.service.ts
@@ -33,7 +33,7 @@ export class HeroesService {
   }
 
   actualizarHeroe(heroe: Heroe): Observable<Heroe> {
-    return this.http.put<Heroe>(`${this.baseUrl}/heroes/${heroe.id}`, heroe); //post actualizar 
+    return this.http.put<Heroe>(`${this.baseUrl}/heroes/${heroe.id}`, heroe); //put actualizar 
   }
 
   borrarHeroe(id: string): Observable<any> {